rest_service/requirements.txt
Django>=2.0,<3.0
djangorestframework
markdown
django-cors-headers
